requirements.txt
streamlit
scapy
pandas
matplotlib
flask
streamlit
speedtest-cli